Changement de double en float

+++ Vector3D.cpp
@@ -5,22 +5,22 @@
 
 // Getters 
 
-double Vector3D::getX() const { return x; }
-double Vector3D::getY() const { return y; }
-double Vector3D::getZ() const { return z; }
+float Vector3D::getX() const { return x; }
+float Vector3D::getY() const { return y; }
+float Vector3D::getZ() const { return z; }
 
 // Public methods
 
-double Vector3D::norm() const {
+float Vector3D::norm() const {
 	return sqrt(this->squaredNorm());
 }
 
-double Vector3D::squaredNorm() const {
+float Vector3D::squaredNorm() const {
 	return x * x + y * y + z * z;
 }
 
 Vector3D Vector3D::normalize() const {
-	double norm = this->norm();
+	float norm = this->norm();
 	return *this / norm;
 }
 
@@ -28,7 +28,7 @@ Vector3D Vector3D::elementWiseProduct(const Vector3D& v) const {
 	return Vector3D(x * v.getX(), y * v.getY(), z * v.getZ());
 }
 
-double Vector3D::scalarProduct(const Vector3D& v) const {
+float Vector3D::scalarProduct(const Vector3D& v) const {
 	return x * v.getX() + y * v.getY() + z * v.getZ();
 }
 
@@ -42,19 +42,19 @@ Vector3D Vector3D::crossProduct(const Vector3D& v) const {
 
 // Operator overloads
 
-Vector3D operator*(const double& c, const Vector3D& v) {
+Vector3D operator*(const float& c, const Vector3D& v) {
 	return Vector3D(c * v.getX(), c * v.getY(), c * v.getZ());
 }
 
-Vector3D operator*(const Vector3D& v, const double& c) {
+Vector3D operator*(const Vector3D& v, const float& c) {
 	return Vector3D(c * v.getX(), c * v.getY(), c * v.getZ());
 }
 
-double operator*(const Vector3D& v1, const Vector3D& v2) {
+float operator*(const Vector3D& v1, const Vector3D& v2) {
 	return v1.scalarProduct(v2);
 }
 
-Vector3D operator/(const Vector3D& v, const double& c) {
+Vector3D operator/(const Vector3D& v, const float& c) {
 	return Vector3D(v.getX() / c, v.getY() / c, v.getZ() / c);
 }
 
@@ -78,7 +78,7 @@ Vector3D::Vector3D() {
 	this->z = 0.0;
 }
 
-Vector3D::Vector3D(double x, double y, double z) {
+Vector3D::Vector3D(float x, float y, float z) {
 	this->x = x;
 	this->y = y;
 	this->z = z;
